Add dynamic label creation from name list
