+++ prac_08/dynamic_labels.kv
@@ -1,0 +1,5 @@
+BoxLayout:
+    orientation: 'vertical'
+    BoxLayout:
+        orientation: 'vertical'
+        id: main
